Slowed the SPI operations down.
This seems to make the LCD operate more consistently.

+++ HAL_MSP432_320x240_ILI9341.h
@@ -15,7 +15,7 @@
 void HAL_LCD_SpiInit();
 void writeData(uint8_t data);
 void writeCommand(uint8_t command);
-void delay(uint16_t msec);
+void delay(uint8_t msec);
 #define LCD_EUSCI_MODULE                      EUSCI_B0_BASE
 // clock P1.1
 #ifndef LCD_SCLK_PIN
